requirements.txt
attrs==22.1.0
click==8.1.3
exceptiongroup==1.0.4
gitdb==4.0.9
GitPython==3.1.29
iniconfig==1.1.1
loguru==0.6.0
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
smmap==5.0.0
tomli==2.0.1
